Refactor emulator jump table
Instead of decoding and executing instructions through multiple
jump tables which incrementally resolve the instruction to the correct
function, use a single jump table, where the keys are formed by building
a 11 bit decode value from funct7, funct3, and opcode. There are some
instructions which don't care about funct7 and/or funct3, in such cases
all the bit variations have to be added as keys, which means longer
jump table, but faster execution, as no further processing is needed.
It can be simplified by creating another table which stores bit masks for
opcodes, and those bit masks are used to get the relevant part of the 11
decode bits, but it adds some overhead because of the extra lookup and masking.