Fix sequential logic and make assignments non-blocking
CPU modules were implemented as combinational always blocks,
but they need to be sequential. To make them sequential, the always
blocks has been changed to be triggered by the positive edge of the
clock. A lot of assignments were defined as blocking, but they
were not dependent on eachother, so most of them has been refactored
to be non-blocking.