Fix synthesis issues and board constraints
rd_valid decoding was wrong, as it used output registers to determine
if rd is valid for the given instruction. For some reason it only
manifested in Vivado simulation and on board, but not in icarus verilog.
Fix board voltage constraints.
Refactor uart always block to be triggered by clock instead of
sample_tick signal.